requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
altgraph==0.17
apiclient==1.0.4
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
b0mb3r==3.0.1.1
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
chromedriver==2.24.1
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
cssselect==1.1.0
DAWG-Python==0.7.2
decorator==5.0.7
discord.py==1.6.0
distlib==0.3.1
Django==3.1.1
docopt==0.6.2
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastapi==0.58.0
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
geocoder==1.38.1
google-api-core==1.17.0
google-api-python-client==1.9.1
google-auth==1.16.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
greenlet==1.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.6.16
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
hupper==1.10.2
hyperframe==5.2.0
idna==2.9
importlib-metadata==3.4.0
iqsms==0.1.1
iqsms-rest==0.0.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
loguru==0.5.1
lxml==4.5.0
MarkupSafe==1.1.1
MouseInfo==0.1.3
multidict==4.7.6
numpy==1.19.2
oauth2client==4.1.3
opencv-python==4.5.1.48
openpyxl==3.0.3
parse==1.15.0
PasteDeploy==2.1.0
pefile==2019.4.18
phonenumbers==8.12.5
Pillow==7.1.1
plaster==1.0
plaster-pastedeploy==0.7
player==0.6.1
prettytable==0.7.2
protobuf==3.12.2
py-vkontakte==5.70.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pydantic==1.5.1
pyee==7.0.1
pygame==2.0.0
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyMsgBox==1.0.9
pyperclip==1.8.2
pyppeteer==0.2.2
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.1.7.3
pyqtgraph==0.11.0
pyquery==1.4.1
pyramid==1.10.4
pyramid-jinja2==2.8
PyRect==0.1.4
PyScreeze==0.1.26
python-docx==0.8.10
python-dotenv==0.14.0
python-telegram-bot==13.5
PyTweening==1.0.3
pytz==2020.1
pywin32-ctypes==0.2.0
ratelim==0.1.6
requests==2.23.0
requests-html==0.10.0
rfc3986==1.4.0
rsa==4.0
selenium==3.141.0
sentry-sdk==0.15.1
six==1.15.0
sniffio==1.1.0
soupsieve==2.0
SQLAlchemy==1.4.5
sqlparse==0.3.1
starlette==0.13.4
tornado==6.1
tqdm==4.46.0
translationstring==1.4
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
uvicorn==0.11.5
venusian==3.0.0
virtualenv==20.4.4
vk-api==11.8.0
vklancer==1.4.0
w3lib==1.21.0
webdriver-manager==3.2.2
WebOb==1.8.6
websockets==8.1
Werkzeug==1.0.1
win32-setctime==1.0.1
WTForms==2.3.3
XlsxWriter==1.2.9
yarl==1.4.2
zipp==3.4.0
zope.deprecation==4.4.0
zope.interface==5.1.0
